Sample JDeveloper workspace to show the issue ADFEMG-108 : ADF Show Printable Page Behavior  Tag  doesn't show printable page in a new wind
ADF Show Printable Page Behavior  Tag  doesn't show printable page in a
new window.
The page rendered in a new window (or tab) shows a default layout insted
of printable one.

Only the second click on a button in a
new window renders printable page layout in it.

Steps to reproduce:

1. Open a sample Workspace attached
2. Run home.jsf in JDeveloper
3. Click on a button print
4a. Doesn't work like expected in IE 8, IE 7, Firefox 3.6.6. -> the same
page is displaed in a new window. The button Print is still visible.
4b. Works like expected in Chrome (24.0.1312.57 m) and Firefox (10.0.12)
-> printable page is displayed in a new window. The button Print is not
visible.